follow and unfollow buttons working

+++ app/helpers/users_helper.rb
@@ -1,14 +1,20 @@
 module UsersHelper
-
   def follow_link (user)
-    link_to "Follow", user_follow_path(user), :method => :post
+    if current_user.following? user
 
+      link_to "Unfollow", user_follow_path(user), :method => :delete
+
+    else
+      link_to "Follow", user_follow_path(user), :method => :post
+
+    end
   end
-
-
-  def unfollow_link (user)
-    link_to "Unfollow", user_follow_path(user), :method => :delete
-
-  end
-
 end
+
+
+#
+# def unfollow_link (user)
+#   link_to "Unfollow", user_follow_path(user), :method => :delete
+#
+# end
+
